Get rid of two warnings "array subscript has type 'char'" caused by passing chars instead of ints to tolower()

+++ src/util.c
@@ -672,9 +672,9 @@ util_strcasestr(const char *haystack, const char *needle) {
   if (needle == NULL || ! *needle )
     return (char *) haystack;
   for (; *haystack; haystack++) {
-    if (tolower(*haystack) == tolower(*needle)) {
+    if (tolower((int)*haystack) == tolower((int)*needle)) {
       const char *h = haystack, *n = needle;
-      for (; *h && *n && tolower(*h) == tolower(*n); ++h, ++n);
+      for (; *h && *n && tolower((int)*h) == tolower((int)*n); ++h, ++n);
       if (! *n ) 
         return (char*)haystack;
     }
